+++ src/tiles/EmptyTile.java
@@ -1,0 +1,13 @@
+package tiles;
+
+public class EmptyTile extends Tile {
+
+    public EmptyTile(int row, int col) {
+        super(row, col);
+    }
+
+    @Override
+    public String getNumberOnTile() {
+        return null;
+    }
+}
